+++ dependency-packager/database/schema/index.js
@@ -1,0 +1,3 @@
+const package = require('./package.js');
+
+module.exports = { package };
