+++ dependency-packager/database/schema/package.js
@@ -1,0 +1,10 @@
+const mongoose = require('mongoose');
+const Schema = mongoose.Schema;
+
+const package = new Schema({
+	module: String,
+	version: String,
+	data: String
+});
+
+module.exports = package;
