+++ dependency-packager/packager/index.js
@@ -1,0 +1,3 @@
+const packager = require('./packager');
+
+module.exports = packager;
