+++ dependency-packager/packager/package-template.js
@@ -1,0 +1,9 @@
+const packageTemplate = `{
+    "name": "dependency",
+    "version": "1.0.0",
+    "main": "index.js",
+    "license": "MIT"
+  }
+  `;
+
+module.exports = packageTemplate;
